feat: 이벤트 카드 mockData 연결 (#74)
